+++ src/extensions/building/buildingext_hooks.cpp
@@ -57,6 +57,7 @@
 #include "iomap.h"
 #include "spritecollection.h"
 #include "extension.h"
+#include "sideext.h"
 #include "fatal.h"
 #include "asserthandler.h"
 #include "debughandler.h"
@@ -77,6 +78,8 @@ static class BuildingClassExt final : public BuildingClass
 public:
     bool _Can_Have_Rally_Point();
     void _Update_Buildables();
+    const InfantryTypeClass* _Crew_Type() const;
+    int _How_Many_Survivors() const;
 };
 
 
@@ -158,6 +161,47 @@ void BuildingClassExt::_Update_Buildables()
             break;
         }
     }
+}
+
+
+/**
+ *  Fetches the kind of crew this object contains.
+ *
+ *  @author: ZivDero
+ */
+const InfantryTypeClass* BuildingClassExt::_Crew_Type() const
+{
+    /**
+     *  Construction yards can sometimes have an engineer exit them.
+     */
+    const int engineer_chance = Extension::Fetch<BuildingTypeClassExtension>(Class)->EngineerChance;
+    if (!IsCaptured && Percent_Chance(engineer_chance))
+        return SideClassExtension::Get_Engineer(House);
+
+    return TechnoClass::Crew_Type();
+}
+
+
+/**
+ *  This determines the maximum number of survivors.
+ *
+ *  @author: 08/04/1996 JLB - Created
+ *           ZivDero - Adjustments for Tiberian Sun
+ */
+int BuildingClassExt::_How_Many_Survivors() const
+{
+    if (IsSurvivorless || !Class->IsCrew)
+        return 0;
+
+    int divisor = SideClassExtension::Get_Survivor_Divisor(House);
+    if (divisor == 0)
+        return 0;
+
+    if (IsCaptured)
+        divisor *= 2;
+
+    const int count = (Class->Cost_Of(House) * Rule->SurvivorFraction) / divisor;
+    return std::clamp(count, 1, 5);
 }
 
 
@@ -796,6 +840,101 @@ DECLARE_PATCH(_BuildingClass_Receive_Message_Only_Allow_Dockable_Harvester_Patch
 
 
 /**
+ *  #issue-445
+ *
+ *  Fixes a bug where crew wouldn't come out of sold/destroyed construction yards
+ *  (or buildings that undeploy, to be more specific).
+ *
+ *  @author: ZivDero
+ */
+DECLARE_PATCH(_BuildingClass_Mission_Deconstruction_ConYard_Survivors_Patch)
+{
+    GET_REGISTER_STATIC(BuildingClass*, this_ptr, esi);
+
+    // This used to be || in RA and is || in YR, but is && in TS, for some reason
+    if (!Target_Legal(this_ptr->ArchiveTarget) || !this_ptr->Class->UndeploysInto)
+    {
+        // Process crew
+        JMP(0x00430CE4);
+    }
+
+    // Don't process crew
+    JMP(0x00430EEA);
+}
+
+
+/**
+ *  Fixes a bug where if when undeploying a construction yard an
+ *  MCV couldn't be placed, it would stay limboed.
+ *
+ *  @author: ZivDero
+ */
+DECLARE_PATCH(_BuildingClass_Mission_Deconstruction_ConYard_Unlimbo_Patch)
+{
+    GET_REGISTER_STATIC(UnitClass*, mcv, ebp);
+    LEA_STACK_STATIC(Coordinate*, coords, esp, 0x40);
+    GET_REGISTER_STATIC(DirType, dir, eax);
+
+    static bool result;
+
+    ScenarioInit++;
+    result = mcv->Unlimbo(*coords, dir);
+    ScenarioInit--;
+
+    if (result)
+    {
+        JMP(0x00430A1A);
+    }
+    else
+    {
+        delete mcv;
+        JMP(0x00430B37);
+    }
+}
+
+
+/**
+ *  Fixes a bug where you could receive double the amount of survivors
+ *  if a building that was being sold got destroyed,
+ *  or free survivors by undeploying a building that was being sold.
+ *
+ *  @author: ZivDero
+ */
+DECLARE_PATCH(_BuildingClass_Mission_Deconstruction_Double_Survivors_Patch)
+{
+    GET_REGISTER_STATIC(BuildingClass*, this_ptr, esi);
+
+    // We've already ejected the survivors, don't eject them any more.
+    this_ptr->IsSurvivorless = true;
+
+    // Stolen instructions
+    this_ptr->Status = 2;
+    this_ptr->Begin_Mode(BSTATE_CONSTRUCTION);
+
+    JMP(0x00430F3B);
+}
+
+
+/**
+ *  Patch to not assign archive targets to buildings currently being sold.
+ *
+ *  @author: ZivDero
+ */
+DECLARE_PATCH(_EventClass_Execute_Archive_Selling_Patch)
+{
+    GET_REGISTER_STATIC(TechnoClass*, techno, edi);
+    GET_REGISTER_STATIC(TARGET, target, eax);
+
+    // Don't assign an archive target if currently selling
+    if (techno->Mission != MISSION_DECONSTRUCTION) {
+        techno->Assign_Archive_Target(target);
+    }
+
+    JMP(0x00494372);
+}
+
+
+/**
  *  Main function for patching the hooks.
  */
 void BuildingClassExtension_Hooks()
@@ -820,4 +959,10 @@ void BuildingClassExtension_Hooks()
     Patch_Jump(0x00426A7E, &_BuildingClass_Receive_Message_Only_Allow_Dockable_Harvester_Patch);
     Patch_Jump(0x00439D10, &BuildingClassExt::_Can_Have_Rally_Point);
     Patch_Jump(0x0042D9A0, &BuildingClassExt::_Update_Buildables);
+    Patch_Jump(0x00433FB0, &BuildingClassExt::_Crew_Type);
+    Patch_Jump(0x00435DA0, &BuildingClassExt::_How_Many_Survivors);
+    Patch_Jump(0x00430CC2, &_BuildingClass_Mission_Deconstruction_ConYard_Survivors_Patch);
+    Patch_Jump(0x00430A01, &_BuildingClass_Mission_Deconstruction_ConYard_Unlimbo_Patch);
+    Patch_Jump(0x00430F2B, &_BuildingClass_Mission_Deconstruction_Double_Survivors_Patch);
+    Patch_Jump(0x0049436A, &_EventClass_Execute_Archive_Selling_Patch);
 }