+++ src/extensions/buildingtype/buildingtypeext.cpp
@@ -50,7 +50,8 @@ BuildingTypeClassExtension::BuildingTypeClassExtension(const BuildingTypeClass *
     ProduceCashBudget(0),
     IsStartupCashOneTime(false),
     IsResetBudgetOnCapture(false),
-    IsEligibleForAllyBuilding(false)
+    IsEligibleForAllyBuilding(false),
+    EngineerChance(0)
 {
     //if (this_ptr) EXT_DEBUG_TRACE("BuildingTypeClassExtension::BuildingTypeClassExtension - Name: %s (0x%08X)\n", Name(), (uintptr_t)(This()));
 
@@ -194,6 +195,7 @@ bool BuildingTypeClassExtension::Read_INI(CCINIClass &ini)
 
     if (!IsInitialized) {
         IsEligibleForAllyBuilding = This()->IsConstructionYard;
+        EngineerChance = This()->ToBuild == RTTI_BUILDINGTYPE ? 25 : 0;
     }
 
     GateUpSound = ini.Get_VocType(ini_name, "GateUpSound", GateUpSound);